# rails_app/swagger/v1/swagger.yaml
...
openapi: 3.0.1
info:
  title: API V1
  version: v1
paths:
...
  "/api/v1/domains":
    get:
      summary: Lists domains for the account
      tags:
        - Domains
      security:
        - bearer_auth: []
      parameters:
        - name: Authorization
          in: header
          required: false
          schema:
            type: string
        - name: X-Signature
          in: header
          required: false
          schema:
            type: string
        - name: X-Timestamp
          in: header
          required: false
          schema:
            type: string
      responses:
        "200":
          description: returns domains for team account after switching
          content:
            application/json:
              schema:
                type: object
                properties:
                  domains:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          description: Unique identifier
                        domain:
                          type: string
                          description: Domain name
                        account_id:
                          type: integer
                          description: Unique identifier
                        is_platform_subdomain:
                          type: boolean
                          description: Whether this is a platform subdomain
                        cloudflare_zone_id:
                          type: string
                          nullable: true
                          description: Cloudflare zone ID
                        dns_verification_status:
                          type: string
                          nullable: true
                          description: DNS verification status
                        dns_last_checked_at:
                          type: string
                          format: date-time
                          nullable: true
                          description: Last DNS check timestamp
                        dns_error_message:
                          type: string
                          nullable: true
                          description: DNS error message
                        website_urls:
                          type: array
                          nullable: true
                          items:
                            type: object
                            properties:
                              id:
                                type: integer
                                description: Unique identifier
                              path:
                                type: string
                                description: URL path
                              website_id:
                                type: integer
                                description: Website ID
                            required:
                              - id
                              - path
                              - website_id
                          description: URLs associated with this domain (only included
                            when include_website_urls=true)
                        created_at:
                          type: string
                          format: date-time
                          description: Timestamp
                        updated_at:
                          type: string
                          format: date-time
                          description: Timestamp
                      required:
                        - id
                        - domain
                        - account_id
                        - is_platform_subdomain
                        - created_at
                        - updated_at
                  platform_subdomain_credits:
                    type: object
                    properties:
                      limit:
                        type: integer
                        description: Maximum platform subdomains allowed
                      used:
                        type: integer
                        description: Number of platform subdomains used
                      remaining:
                        type: integer
                        description: Remaining platform subdomains
                    required:
                      - limit
                      - used
                      - remaining
                  plan_tier:
                    type: string
                    nullable: true
                    description: User plan tier (starter, growth, pro)
                required:
                  - domains
                  - platform_subdomain_credits
        "401":
          description: unauthorized - missing token
    post:
      summary: Creates a domain
      tags:
        - Domains
      security:
        - bearer_auth: []
      parameters:
        - name: Authorization
          in: header
          required: false
          schema:
            type: string
        - name: X-Signature
          in: header
          required: false
          schema:
            type: string
        - name: X-Timestamp
          in: header
          required: false
          schema:
            type: string
      responses:
        "201":
          description: domain created in team account after switching
          content:
            application/json:
              schema:
                type: object
                properties:
                  domain:
                    type: object
                    properties:
                      id:
                        type: integer
                        description: Unique identifier
                      domain:
                        type: string
                        description: Domain name
                      account_id:
                        type: integer
                        description: Unique identifier
                      is_platform_subdomain:
                        type: boolean
                        description: Whether this is a platform subdomain
                      cloudflare_zone_id:
                        type: string
                        nullable: true
                        description: Cloudflare zone ID
                      dns_verification_status:
                        type: string
                        nullable: true
                        description: DNS verification status
                      dns_last_checked_at:
                        type: string
                        format: date-time
                        nullable: true
                        description: Last DNS check timestamp
                      dns_error_message:
                        type: string
                        nullable: true
                        description: DNS error message
                      website_urls:
                        type: array
                        nullable: true
                        items:
                          type: object
                          properties:
                            id:
                              type: integer
                              description: Unique identifier
                            path:
                              type: string
                              description: URL path
                            website_id:
                              type: integer
                              description: Website ID
                          required:
                            - id
                            - path
                            - website_id
                        description: URLs associated with this domain (only included
                          when include_website_urls=true)
                      created_at:
                        type: string
                        format: date-time
                        description: Timestamp
                      updated_at:
                        type: string
                        format: date-time
                        description: Timestamp
                    required:
                      - id
                      - domain
                      - account_id
                      - is_platform_subdomain
                      - created_at
                      - updated_at
                  website_url:
                    type: object
                    properties:
                      id:
                        type: integer
                        description: Unique identifier
                      path:
                        type: string
                        description: URL path
                      account_id:
                        type: integer
                        description: Unique identifier
                      website_id:
                        type: integer
                        description: Unique identifier
                      domain_id:
                        type: integer
                        description: Unique identifier
                      domain_string:
                        type: string
                        nullable: true
                        description: Domain name
                      created_at:
                        type: string
                        format: date-time
                        description: Timestamp
                      updated_at:
                        type: string
                        format: date-time
                        description: Timestamp
                    required:
                      - id
                      - path
                      - account_id
                      - website_id
                      - domain_id
                      - created_at
                      - updated_at
                  platform_subdomain_credits:
                    type: object
                    properties:
                      limit:
                        type: integer
                        description: Maximum platform subdomains allowed
                      used:
                        type: integer
                        description: Number of platform subdomains used
                      remaining:
                        type: integer
                        description: Remaining platform subdomains
                    required:
                      - limit
                      - used
                      - remaining
                required:
                  - domain
                  - website_url
                  - platform_subdomain_credits
        "422":
          description: subdomain limit exceeded
        "401":
          description: unauthorized - missing token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                domain:
                  type: object
                  properties:
                    domain:
                      type: string
                      description: Domain name
                    website_id:
                      type: integer
                      description: Associated website ID
                    is_platform_subdomain:
                      type: boolean
                      description: Whether this is a platform subdomain
              required:
                - domain
...
  "/api/v1/domains/{id}":
    patch:
      summary: Updates domain DNS verification status
      tags:
        - Domains
      security:
        - bearer_auth: []
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
        - name: X-Signature
          in: header
          required: false
          schema:
            type: string
        - name: X-Timestamp
          in: header
          required: false
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: successfully updates dns_verification_status
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: Unique identifier
                  domain:
                    type: string
                    description: Domain name
                  account_id:
                    type: integer
                    description: Unique identifier
                  is_platform_subdomain:
                    type: boolean
                    description: Whether this is a platform subdomain
                  cloudflare_zone_id:
                    type: string
                    nullable: true
                    description: Cloudflare zone ID
                  dns_verification_status:
                    type: string
                    nullable: true
                    description: DNS verification status
                  dns_last_checked_at:
                    type: string
                    format: date-time
                    nullable: true
                    description: Last DNS check timestamp
                  dns_error_message:
                    type: string
                    nullable: true
                    description: DNS error message
                  created_at:
                    type: string
                    format: date-time
                    description: Timestamp
                  updated_at:
                    type: string
                    format: date-time
                    description: Timestamp
                required:
                  - id
                  - domain
                  - account_id
                  - is_platform_subdomain
                  - created_at
                  - updated_at
        "404":
          description: returns not found for non-existent domain
        "401":
          description: unauthorized - missing token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                domain:
                  type: object
                  properties:
                    dns_verification_status:
                      type: string
    parameters:
      - name: id
        in: path
        description: Domain ID
        required: true
        schema:
          type: integer
    get:
      summary: Retrieves a domain
      tags:
        - Domains
      security:
        - bearer_auth: []
      parameters:
        - name: Authorization
          in: header
          required: false
          schema:
            type: string
        - name: X-Signature
          in: header
          required: false
          schema:
            type: string
        - name: X-Timestamp
          in: header
          required: false
          schema:
            type: string
      responses:
        "200":
          description: domain found in team account after switching
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: Unique identifier
                  domain:
                    type: string
                    description: Domain name
                  account_id:
                    type: integer
                    description: Unique identifier
                  is_platform_subdomain:
                    type: boolean
                    description: Whether this is a platform subdomain
                  cloudflare_zone_id:
                    type: string
                    nullable: true
                    description: Cloudflare zone ID
                  dns_verification_status:
                    type: string
                    nullable: true
                    description: DNS verification status
                  dns_last_checked_at:
                    type: string
                    format: date-time
                    nullable: true
                    description: Last DNS check timestamp
                  dns_error_message:
                    type: string
                    nullable: true
                    description: DNS error message
                  website_urls:
                    type: array
                    nullable: true
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          description: Unique identifier
                        path:
                          type: string
                          description: URL path
                        website_id:
                          type: integer
                          description: Website ID
                      required:
                        - id
                        - path
                        - website_id
                    description: URLs associated with this domain (only included when
                      include_website_urls=true)
                  created_at:
                    type: string
                    format: date-time
                    description: Timestamp
                  updated_at:
                    type: string
                    format: date-time
                    description: Timestamp
                required:
                  - id
                  - domain
                  - account_id
                  - is_platform_subdomain
                  - created_at
                  - updated_at
        "404":
          description: cannot access other user domain
...
servers:
  - url: https://{defaultHost}
    variables:
      defaultHost:
        default: www.example.com